% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{backscatter.xsection}
\alias{backscatter.xsection}
\title{Backscattering cross-section of a Scatterer.}
\usage{
backscatter.xsection(scatterer, k, phase.sd = 0, method = c("discrete",
  "continuous"))
}
\arguments{
\item{scatterer}{A Scatterer object.}

\item{k}{A three-element vector, giving the x, y, and z components of the wavenumber
vector corresponding to the incident sound field.}

\item{phase.sd}{The standard deviation of the random phase component added to the form function
of each of the scatterer's segments.  In degrees, defaults to zero. Ignored if method is "continuous".}

\item{method}{The method to use when evaluating the line integral, defaults to "discrete".}
}
\value{
The backscattering cross section of the Scatterer.
}
\description{
Calculates the backscattering cross-section of a scatterer at an arbitrary
frequency and incident angle.
}
\details{
The two required arguments are scatterer and k.  The wavenumber vector has a magnitude
equal to the acoustic wavenumber, given by 2 * pi * f / c, where f is frequency and c is sound speed.
For a normal downward-looking echosounder, this is c(0, 0, -k).

The two available integration methods are "discrete" and "continuous."  The discrete method
corresponds to that used in papers by McGeehee et al. (1998), and the series by Demer and Conti.
The continuous method is an experimental method which, instead of using the discretized outline of the
animal to define a series of short cylinders, uses it to interpolate the positions, radii, and
material properties along the animal's length, and integrate these functions continuously.
}

