% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{frequency.spectrum}
\alias{frequency.spectrum}
\title{Backscattering cross section as a function of frequency.}
\usage{
\method{frequency}{spectrum}(scatterer, freq.start, freq.stop, sound.speed,
  nfreq = 100, ...)
}
\arguments{
\item{scatterer}{A Scatterer object}

\item{freq.start,freq.stop}{The end points of the frequency range to calculate.}

\item{sound.speed}{The speed of sound in the surrounding water.}

\item{nfreq}{The number of frequencies between freq.start and freq.stop.  Defualts
to 100.}

\item{...}{Additional arguments passed to backscatter.xsection.}
}
\value{
A list with elements freqs, sigma.bs, and TS, containing the frequencies
and their corresponding backscatter cross-sections and target strengths.
}
\description{
Backscattering cross section as a function of frequency.
}

