% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{load.scatterer}
\alias{load.scatterer}
\title{Load a Scatterer from a CSV file.}
\usage{
load.scatterer(filename, ...)
}
\arguments{
\item{filename}{Path to the file}

\item{...}{Additional arguments passed to read.csv}
}
\description{
Reads in a scatterer from a CSV file with columns matching the arguments to
Scatterer().
}

