% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert between degrees and radians.}
\usage{
rad2deg <- function(x) x * 180 / pi
deg2rad <- function(x) x * pi / 180
}
\arguments{
\item{x}{Angle to convert}
}
\description{
Convert between degrees and radians.
}
\examples{
rad2deg(pi)
[1] 180
deg2rad(180)
[1] 3.141593
}

