% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{rotate.Scatterer}
\alias{rotate.Scatterer}
\title{Rotate a Scatterer in space.}
\usage{
\method{rotate}{Scatterer}(scatterer, roll = 0, tilt = 0, yaw = 0)
}
\arguments{
\item{scatterer}{A Scatterer object}

\item{roll}{Angle to roll the scatterer, in degrees. Defaults to 0.}

\item{tilt}{Angle to tilt the scatterer, in degrees. Defaults to 0.}

\item{yaw}{Angle to yaw the scatterer, in degrees. Defaults to 0.}
}
\value{
A Scatterer with the same shape and properties, but a new orientation.
}
\description{
Tilts and/or rotates a scatterer, returning a new scatterer object.
}
\details{
The roll, tilt, and yaw refer to rotations around the x, y, and z axes,
respectively. They are applied in that order.
}

