% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{save.scatterer.Scatterer}
\alias{save.scatterer.Scatterer}
\title{Save a Scatterer object to file.}
\usage{
\method{save.scatterer}{Scatterer}(scatterer, filename, ...)
}
\arguments{
\item{scatterer}{The Scatterer object to save}

\item{filename}{Filename to save to}

\item{...}{Additional arguments passed to write.csv}
}
\description{
Writes a Scatterer object to a CSV file.
}

