% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{Scatterer}
\alias{Scatterer}
\title{Scatterer Object}
\usage{
Scatterer(x, y, z, a, g, h)
}
\arguments{
\item{x,y,z}{Vectors defining the scatterer's centerline.  x and y are horizontal, z is up.}

\item{a}{Vector of radii}

\item{g}{Vector of density contrasts}

\item{h}{Vector of sound-speed contrasts}
}
\description{
This function creates Scatterer objects, which encapsulate the shape and
material properties of a fluid-like object with a distorted cylindrical shape
which scatters sound.  The shape is approximated by a series of short cylinders,
which can have varying radii, densities, and sound speeds.
}

