% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{target.strength}
\alias{target.strength}
\title{Target Strength of a Scatterer.}
\usage{
target.strength(scatterer, k, phase.sd = 0, method = c("discrete",
  "continuous"))
}
\arguments{
\item{scatterer}{A Scatterer object.}

\item{k}{A three-element vector, giving the x, y, and z components of the wavenumber
vector corresponding to the incident sound field.}

\item{phase.sd}{The standard deviation of the random phase component added to the form function
of each of the scatterer's segments.  In degrees, defaults to zero. Ignored if method is "continuous".}

\item{method}{The method to use when evaluating the line integral, defaults to "discrete".}
}
\value{
The target strength (i.e., 10 * log10(backscatter.xsection(scatterer, k)))
}
\description{
This function is a simple wrapper around backscatter.xsection.
}
\seealso{
backscatter.xsection
}

