% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdwba.R
\name{tilt.spectrum}
\alias{tilt.spectrum}
\title{Backscattering cross-section as a function of tilt angle.}
\usage{
tilt.spectrum(scatterer, angle.start, angle.stop, freq, sound.speed,
  nangle = 100, ...)
}
\arguments{
\item{scatterer}{A Scatterer object}

\item{angle.start,angle.stop}{The end points of the angle range to calculate.}

\item{sound.speed}{The speed of sound in the surrounding water.}

\item{nangle}{The number of frequencies between angle.start and angle.stop.  Defualts
to 100.}

\item{...}{Additional arguments passed to backscatter.xsection.}
}
\description{
Backscattering cross-section as a function of tilt angle.
}

